add inheritance structure for network
