+++ pypsa/components/abstract.py
@@ -1,9 +1,8 @@
 """
 Abstract components module.
 
-Contains classes and properties relevant to all component types in PyPSA. Also imports
-logic from other modules:
-- components.types
+Only defines a base class for all Components helper classes which inherit to
+`Components` class.
 """
 
 from __future__ import annotations
