+++ pypsa/network/__init__.py
@@ -1,0 +1,1 @@
+"""Package for network specific functionality in PyPSA."""
