+++ pypsa/networks.py
@@ -52,7 +52,6 @@ from pypsa.descriptors import (
 )
 from pypsa.graph import adjacency_matrix, graph, incidence_matrix
 from pypsa.io import (
-    _import_components_from_df,
     _import_series_from_df,
     export_to_csv_folder,
     export_to_excel,
@@ -68,6 +67,8 @@ from pypsa.io import (
     import_series_from_dataframe,
     merge,
 )
+from pypsa.network.index import _NetworkIndex
+from pypsa.network.transform import _NetworkTransform
 from pypsa.optimization.optimize import OptimizationAccessor
 from pypsa.pf import (
     calculate_B_H,
@@ -85,7 +86,6 @@ from pypsa.pf import (
 from pypsa.plot.accessor import PlotAccessor
 from pypsa.plot.maps import explore, iplot
 from pypsa.statistics import StatisticsAccessor
-from pypsa.typing import is_1d_list_like
 from pypsa.version import __version_semver__
 
 if TYPE_CHECKING:
@@ -117,7 +117,7 @@ def _create_component_property(property_type: str, component: str) -> property:
     return property(getter, setter)
 
 
-class Network:
+class Network(_NetworkTransform, _NetworkIndex):
     """Network container for all buses, one-ports and branches."""
 
     # Type hints
@@ -786,1001 +786,6 @@ class Network:
 
         """
         self.crs = pyproj.CRS.from_epsg(new)
-
-    # ----------------
-    # Indexers
-    # ----------------
-
-    def set_snapshots(
-        self,
-        snapshots: Sequence,
-        default_snapshot_weightings: float = 1.0,
-        weightings_from_timedelta: bool = False,
-    ) -> None:
-        """
-        Set the snapshots/time steps and reindex all time-dependent data.
-
-        Snapshot weightings, typically representing the hourly length of each snapshot,
-        is filled with the `default_snapshot_weighintgs` value, or uses the timedelta
-        of the snapshots if `weightings_from_timedelta` flag is True, and snapshots are
-        of type `pd.DatetimeIndex`.
-
-        This will reindex all components time-dependent DataFrames
-        (:py:meth:`pypsa.Network.dynamic`). NaNs are filled with the default value for
-        that quantity.
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.set_snapshots(pd.date_range("2015-01-01", freq="h", periods=3))
-        >>> n.snapshots
-        DatetimeIndex(['2015-01-01 00:00:00', '2015-01-01 01:00:00',
-                       '2015-01-01 02:00:00'],
-                      dtype='datetime64[ns]', name='snapshot', freq='h')
-
-        Parameters
-        ----------
-        snapshots : list, pandas.Index or pd.MultiIndex
-            All time steps.
-        default_snapshot_weightings: float
-            The default weight for each snapshot. Defaults to 1.0.
-        weightings_from_timedelta: bool
-            Wheter to use the timedelta of `snapshots` as `snapshot_weightings` if
-            `snapshots` is of type `pd.DatetimeIndex`.  Defaults to False.
-
-        Returns
-        -------
-        None
-
-        """
-        # Check if snapshots contain timezones
-        if isinstance(snapshots, pd.DatetimeIndex) and snapshots.tz is not None:
-            msg = (
-                "Numpy datetime64[ns] objects with timezones are not supported and are "
-                "thus not allowed in snapshots. Please pass timezone-naive timestamps "
-                "(e.g. via ds.values)."
-            )
-            raise ValueError(msg)
-
-        if isinstance(snapshots, pd.MultiIndex):
-            if snapshots.nlevels != 2:
-                msg = "Maximally two levels of MultiIndex supported"
-                raise ValueError(msg)
-            sns = snapshots.rename(["period", "timestep"])
-            sns.name = "snapshot"
-            self._snapshots = sns
-        else:
-            self._snapshots = pd.Index(snapshots, name="snapshot")
-
-        if len(self._snapshots) == 0:
-            msg = "Snapshots must not be empty."
-            raise ValueError(msg)
-
-        self.snapshot_weightings = self.snapshot_weightings.reindex(
-            self._snapshots, fill_value=default_snapshot_weightings
-        )
-
-        if isinstance(snapshots, pd.DatetimeIndex) and weightings_from_timedelta:
-            hours_per_step = (
-                snapshots.to_series()
-                .diff(periods=1)
-                .shift(-1)
-                .ffill()  # fill last value by assuming same as the one before
-                .apply(lambda x: x.total_seconds() / 3600)
-            )
-            self._snapshot_weightings = pd.DataFrame(
-                dict.fromkeys(self._snapshot_weightings.columns, hours_per_step)
-            )
-        elif not isinstance(snapshots, pd.DatetimeIndex) and weightings_from_timedelta:
-            logger.info(
-                "Skipping `weightings_from_timedelta` as `snapshots`is not of type `pd.DatetimeIndex`."
-            )
-
-        for component in self.all_components:
-            dynamic = self.dynamic(component)
-            attrs = self.components[component]["attrs"]
-
-            for k in dynamic:
-                if dynamic[k].empty:  # avoid expensive reindex operation
-                    dynamic[k].index = self._snapshots
-                elif k in attrs.default[attrs.varying]:
-                    if isinstance(dynamic[k].index, pd.MultiIndex):
-                        dynamic[k] = dynamic[k].reindex(
-                            self._snapshots, fill_value=attrs.default[attrs.varying][k]
-                        )
-                    else:
-                        # Make sure to keep timestep level in case of MultiIndex
-                        dynamic[k] = dynamic[k].reindex(
-                            self._snapshots,
-                            fill_value=attrs.default[attrs.varying][k],
-                            level="timestep",
-                        )
-                else:
-                    dynamic[k] = dynamic[k].reindex(self._snapshots)
-
-        # NB: No need to rebind dynamic to self, since haven't changed it
-
-    @property
-    def snapshots(self) -> pd.Index | pd.MultiIndex:
-        """
-        Snapshots dimension of the network.
-
-        If snapshots are a pandas.MultiIndex, the first level are investment periods
-        and the second level are timesteps. If snapshots are single indexed, the only
-        level is timesteps.
-
-        Returns
-        -------
-        pd.Index or pd.MultiIndex
-            Snapshots of the network, either as a single index or a multi-index.
-
-        Examples
-        --------
-        >>> n.snapshots # doctest: +ELLIPSIS
-        DatetimeIndex(['2015-01-01 00:00:00', '2015-01-01 01:00:00',
-                       '2015-01-01 02:00:00', '2015-01-01 03:00:00',
-                      ...
-                      dtype='datetime64[ns]', name='snapshot', freq=None)
-
-        See Also
-        --------
-        pypsa.networks.Network.timesteps : Get the timestep level only.
-        pypsa.networks.Network.periods : Get the period level only.
-
-        Notes
-        -----
-        Note that Snapshots are a dimension, while timesteps and and periods are
-        only levels of the snapshots dimension, similar to coords in xarray.
-        This is because timesteps and periods are not necessarily unique or complete
-        across snapshots.
-
-        """
-        return self._snapshots
-
-    @snapshots.setter
-    def snapshots(self, snapshots: Sequence) -> None:
-        """
-        Setter for snapshots dimension.
-
-        Parameters
-        ----------
-        snapshots : Sequence
-            Snapshots to be set.
-
-        See Also
-        --------
-        pypsa.networks.Network.snapshots : Getter method
-        pypsa.networks.Network.set_snapshots : Setter method
-
-        """
-        self.set_snapshots(snapshots)
-
-    @property
-    def timesteps(self) -> pd.Index:
-        """
-        Timestep level of snapshots dimension.
-
-        If snapshots is single indexed, timesteps and snapshots yield the same result.
-        Otherwise only the timestep level will be returned.
-
-        Returns
-        -------
-        pd.Index
-            Timesteps of the network.
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.set_snapshots(pd.date_range("2015-01-01", freq="h", periods=3))
-
-        For a Network without investment periods, the timesteps are identical to the
-        snapshots:
-        >>> n.timesteps
-        DatetimeIndex(['2015-01-01 00:00:00', '2015-01-01 01:00:00',
-                       '2015-01-01 02:00:00'],
-                      dtype='datetime64[ns]', name='snapshot', freq='h')
-        >>> n.snapshots
-        DatetimeIndex(['2015-01-01 00:00:00', '2015-01-01 01:00:00',
-                       '2015-01-01 02:00:00'],
-                      dtype='datetime64[ns]', name='snapshot', freq='h')
-
-        For a Network with investment periods, the timesteps are are the unqiue set
-        of timesteps in across all investment periods:
-        >>> n.investment_periods = [1, 2]
-        >>> n.timesteps
-        DatetimeIndex(['2015-01-01 00:00:00', '2015-01-01 01:00:00',
-                       '2015-01-01 02:00:00'],
-                      dtype='datetime64[ns]', name='timestep', freq=None)
-        >>> n.snapshots
-        MultiIndex([(1, '2015-01-01 00:00:00'),
-                (1, '2015-01-01 01:00:00'),
-                (1, '2015-01-01 02:00:00'),
-                (2, '2015-01-01 00:00:00'),
-                (2, '2015-01-01 01:00:00'),
-                (2, '2015-01-01 02:00:00')],
-               name='snapshot')
-
-        See Also
-        --------
-        pypsa.networks.Network.snapshots : Get the snapshots dimension.
-        pypsa.networks.Network.periods : Get the period level only.
-
-        """
-        if "timestep" in self.snapshots.names:
-            return self.snapshots.get_level_values("timestep").unique()
-        return self.snapshots
-
-    @timesteps.setter
-    def timesteps(self, timesteps: Sequence) -> None:
-        """
-        Setter for timesteps level of snapshots dimension.
-
-        .. warning::
-            Setting `timesteps` is not supported. Please set `snapshots` instead.
-
-        Parameters
-        ----------
-        timesteps : Sequence
-            Timesteps to be set.
-
-        Also see
-        --------
-        pypsa.networks.Network.timesteps : Getter method
-
-        """
-        msg = "Setting `timesteps` is not supported. Please set `snapshots` instead."
-        raise NotImplementedError(msg)
-
-    @property
-    def periods(self) -> pd.Index:
-        """
-        Periods level of snapshots dimension.
-
-        If snapshots is single indexed, periods will always be empty, since there no
-        investment periods without timesteps are defined. Otherwise only the period
-        level will be returned.
-
-        Returns
-        -------
-        pd.Index
-            Periods of the network.
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", "bus") # doctest: +SKIP
-        >>> n.snapshots = pd.date_range("2015-01-01", freq="h", periods=3)
-        >>> n.snapshots
-        DatetimeIndex(['2015-01-01 00:00:00', '2015-01-01 01:00:00',
-                       '2015-01-01 02:00:00'],
-                      dtype='datetime64[ns]', name='snapshot', freq='h')
-
-        Add investment periods:
-        >>> n.periods = [1, 2]
-        >>> n.periods
-        Index([1, 2], dtype='int64', name='period')
-
-        Which will also cast snapshots to a MultiIndex:
-        >>> n.snapshots
-        MultiIndex([(1, '2015-01-01 00:00:00'),
-                (1, '2015-01-01 01:00:00'),
-                (1, '2015-01-01 02:00:00'),
-                (2, '2015-01-01 00:00:00'),
-                (2, '2015-01-01 01:00:00'),
-                (2, '2015-01-01 02:00:00')],
-               name='snapshot')
-
-        See Also
-        --------
-        pypsa.networks.Network.snapshots : Get the snapshots dimension.
-        pypsa.networks.Network.timesteps : Get the timestep level only.
-
-        """
-        if "period" in self.snapshots.names:
-            return self.snapshots.get_level_values("period").unique()
-        return pd.Index([], name="period")
-
-    @periods.setter
-    def periods(self, periods: Sequence) -> None:
-        """
-        Setter for periods level of snapshots dimension.
-
-        Parameters
-        ----------
-        periods : Sequence
-            Investment periods to be set.
-        Also see
-        --------
-        pypsa.networks.Network.periods : Getter method
-        pypsa.networks.Network.set_investment_periods : Setter method
-
-        """
-        self.set_investment_periods(periods)
-
-    @property
-    def has_periods(self) -> bool:
-        """
-        Check if network has investment periods assigned to snapshots dimension.
-
-        Returns
-        -------
-        bool
-            True if network has investment periods, otherwise False.
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", "bus") # doctest: +SKIP
-        >>> n.snapshots = pd.date_range("2015-01-01", freq="h", periods=3)
-        >>> n.has_periods
-        False
-
-        Add investment periods:
-        >>> n.periods = [1, 2]
-        >>> n.has_periods
-        True
-
-        See Also
-        --------
-        pypsa.networks.Network.snapshots : Snapshots dimension of the network.
-
-        """
-        return not self.periods.empty
-
-    @property
-    def investment_periods(self) -> pd.Index:
-        """
-        Periods level of snapshots dimension.
-
-        If snapshots is single indexed, periods will always be empty, since there no
-        investment periods without timesteps are defined. Otherwise only the period
-        level will be returned.
-
-        .. Note :: Alias for :py:meth:`pypsa.Network.periods`.
-
-        Returns
-        -------
-        pd.Index
-            Investment periods of the network.
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", "bus") # doctest: +SKIP
-        >>> n.snapshots = pd.date_range("2015-01-01", freq="h", periods=3)
-        >>> n.snapshots
-        DatetimeIndex(['2015-01-01 00:00:00', '2015-01-01 01:00:00',
-                       '2015-01-01 02:00:00'],
-                      dtype='datetime64[ns]', name='snapshot', freq='h')
-
-        Add investment periods:
-        >>> n.investment_periods = [1, 2]
-        >>> n.investment_periods
-        Index([1, 2], dtype='int64', name='period')
-
-        Which will also cast snapshots to a MultiIndex:
-        >>> n.snapshots
-        MultiIndex([(1, '2015-01-01 00:00:00'),
-                (1, '2015-01-01 01:00:00'),
-                (1, '2015-01-01 02:00:00'),
-                (2, '2015-01-01 00:00:00'),
-                (2, '2015-01-01 01:00:00'),
-                (2, '2015-01-01 02:00:00')],
-               name='snapshot')
-
-        See Also
-        --------
-        pypsa.networks.Network.snapshots : Get the snapshots dimension.
-        pypsa.networks.Network.periods : Get the snapshots dimension.
-        pypsa.networks.Network.timesteps : Get the timestep level only.
-
-        """
-        return self.periods
-
-    @investment_periods.setter
-    def investment_periods(self, periods: Sequence) -> None:
-        """
-        Setter for periods level of snapshots dimension.
-
-        .. Note :: Alias for :py:meth:`pypsa.Network.periods`.
-
-        Parameters
-        ----------
-        periods : Sequence
-            Investment periods to be set.
-
-        Also see
-        --------
-        pypsa.networks.Network.periods : Getter method
-        pypsa.networks.Network.set_investment_periods : Setter method
-
-        """
-        self.periods = periods
-
-    @property
-    def has_investment_periods(self) -> bool:
-        """
-        Check if network has investment periods assigned to snapshots dimension.
-
-        .. Note :: Alias for :py:meth:`pypsa.Network.has_periods`.
-
-        Returns
-        -------
-        bool
-            True if network has investment periods, otherwise False.
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", "bus") # doctest: +SKIP
-        >>> n.snapshots = pd.date_range("2015-01-01", freq="h", periods=3)
-        >>> n.has_investment_periods
-        False
-
-        Add investment periods:
-        >>> n.periods = [1, 2]
-        >>> n.has_investment_periods
-        True
-
-        See Also
-        --------
-        pypsa.networks.Network.snapshots : Snapshots dimension of the network.
-        pypsa.networks.Network.periods : Periods level of snapshots dimension.
-
-        """
-        return self.has_periods
-
-    @property
-    def snapshot_weightings(self) -> pd.DataFrame:
-        """
-        Weightings applied to each snapshots during the optimization (LOPF).
-
-        * Objective weightings multiply the operational cost in the
-          objective function.
-
-        * Generator weightings multiply the impact of all generators
-          in global constraints, e.g. multiplier of GHG emmissions.
-
-        * Store weightings define the elapsed hours for the charge, discharge
-          standing loss and spillage of storage units and stores in order to
-          determine the state of charge.
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", "bus") # doctest: +SKIP
-        >>> n.snapshots = pd.date_range("2015-01-01", freq="h", periods=3)
-        >>> n.snapshot_weightings
-                         objective  stores  generators
-        snapshot
-        2015-01-01 00:00:00        1.0     1.0         1.0
-        2015-01-01 01:00:00        1.0     1.0         1.0
-        2015-01-01 02:00:00        1.0     1.0         1.0
-
-        Change the snapshot weightings:
-        >>> n.snapshot_weightings.objective = [5, 7, 9]
-        >>> n.snapshot_weightings
-                         objective  stores  generators
-        snapshot
-        2015-01-01 00:00:00          5     1.0         1.0
-        2015-01-01 01:00:00          7     1.0         1.0
-        2015-01-01 02:00:00          9     1.0         1.0
-
-        """
-        return self._snapshot_weightings
-
-    @snapshot_weightings.setter
-    def snapshot_weightings(self, df: pd.DataFrame) -> None:
-        if not df.index.equals(self.snapshots):
-            msg = "Weightings not defined for all snapshots."
-            raise ValueError(msg)
-
-        if isinstance(df, pd.Series):
-            logger.info("Applying weightings to all columns of `snapshot_weightings`")
-            df = pd.DataFrame(dict.fromkeys(self._snapshot_weightings.columns, df))
-        self._snapshot_weightings = df
-
-    def set_investment_periods(self, periods: Sequence) -> None:
-        """
-        Set the investment periods of the network.
-
-        If the network snapshots are a pandas.MultiIndex, the investment periods
-        have to be a subset of the first level. If snapshots are a single index,
-        they and all time-series are repeated for all periods. This changes
-        the network snapshots to be a MultiIndex (inplace operation) with the first
-        level being the investment periods and the second level the snapshots.
-
-        Parameters
-        ----------
-        n : pypsa.Network
-        periods : list
-            List of periods to be selected/initialized.
-
-        Returns
-        -------
-        None.
-
-        """
-        periods_ = pd.Index(periods, name="period")
-        if periods_.empty:
-            return
-        if not (
-            pd.api.types.is_integer_dtype(periods_)
-            and periods_.is_unique
-            and periods_.is_monotonic_increasing
-        ):
-            msg = (
-                "Investment periods are not strictly increasing integers, "
-                "which is required for multi-period investment optimisation."
-            )
-            raise ValueError(msg)
-        if isinstance(self.snapshots, pd.MultiIndex):
-            if not periods_.isin(self.snapshots.unique("period")).all():
-                msg = "Not all investment periods are in level `period` of snapshots."
-                raise ValueError(msg)
-            if len(periods_) < len(self.snapshots.unique(level="period")):
-                msg = "Investment periods do not equal first level values of snapshots."
-                raise NotImplementedError(msg)
-        else:
-            # Convenience case:
-            logger.info(
-                "Repeating time-series for each investment period and "
-                "converting snapshots to a pandas.MultiIndex."
-            )
-            names = ["period", "timestep"]
-            for component in self.all_components:
-                dynamic = self.dynamic(component)
-
-                for k in dynamic:
-                    dynamic[k] = pd.concat(
-                        dict.fromkeys(periods_, dynamic[k]), names=names
-                    )
-                    dynamic[k].index.name = "snapshot"
-
-            self._snapshots = pd.MultiIndex.from_product(
-                [periods_, self.snapshots], names=names
-            )
-            self._snapshots.name = "snapshot"
-            self._snapshot_weightings = pd.concat(
-                dict.fromkeys(periods_, self.snapshot_weightings), names=names
-            )
-            self._snapshot_weightings.index.name = "snapshot"
-
-        self.investment_period_weightings = self.investment_period_weightings.reindex(
-            self.periods, fill_value=1.0
-        ).astype(float)
-
-    @property
-    def investment_period_weightings(self) -> pd.DataFrame:
-        """
-        Weightings applied to each investment period during the optimization (LOPF).
-
-        Objective weightings are multiplied with all cost coefficients in the
-        objective function of the respective investment period (e.g. to include a
-        social discount rate).
-        Years weightings denote the elapsed time until the subsequent investment period
-        (e.g. used for global constraints CO2 emissions).
-
-        Examples
-        --------
-        Create a network with investment periods:
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", "bus") # doctest: +SKIP
-        >>> n.snapshots = pd.date_range("2015-01-01", freq="h", periods=2)
-        >>> n.investment_periods = [1, 2]
-
-        >>> n.investment_period_weightings
-                objective  years
-        period
-        1             1.0    1.0
-        2             1.0    1.0
-
-        Change the investment period weightings:
-        >>> n.investment_period_weightings.objective = [5, 7]
-        >>> n.investment_period_weightings.years = [1, 2]
-        >>> n.investment_period_weightings
-                objective  years
-        period
-        1             5    1
-        2             7    2
-
-        """
-        return self._investment_period_weightings
-
-    @investment_period_weightings.setter
-    def investment_period_weightings(self, df: pd.DataFrame) -> None:
-        if not df.index.equals(self.investment_periods):
-            msg = "Weightings not defined for all investment periods."
-            raise ValueError(msg)
-        if isinstance(df, pd.Series):
-            logger.info(
-                "Applying weightings to all columns of `investment_period_weightings`"
-            )
-            df = pd.DataFrame(
-                dict.fromkeys(self._investment_period_weightings.columns, df)
-            )
-        self._investment_period_weightings = df
-
-    def add(
-        self,
-        class_name: str,
-        name: str | int | Sequence[int | str],
-        suffix: str = "",
-        overwrite: bool = False,
-        **kwargs: Any,
-    ) -> pd.Index:
-        """
-        Add components to the network.
-
-        Handles addition of single and multiple components along with their attributes.
-        Pass a list of names to add multiple components at once or pass a single name
-        to add a single component.
-
-        When a single component is added, all non-scalar attributes are assumed to be
-        time-varying and indexed by snapshots.
-        When multiple components are added, all non-scalar attributes are assumed to be
-        static and indexed by names. A single value sequence is treated as scalar and
-        broadcasted to all components. It is recommended to explicitly pass a scalar
-        instead.
-        If you want to add time-varying attributes to multiple components, you can pass
-        a 2D array/ DataFrame where the first dimension is snapshots and the second
-        dimension is names.
-
-        Any attributes which are not specified will be given the default
-        value from :doc:`/user-guide/components`.
-
-        Parameters
-        ----------
-        class_name : str
-            Component class name in ("Bus", "Generator", "Load", "StorageUnit",
-            "Store", "ShuntImpedance", "Line", "Transformer", "Link").
-        name : str or int or list of str or list of int
-            Component name(s)
-        suffix : str, default ""
-            All components are named after name with this added suffix.
-        overwrite : bool, default False
-            If True, existing components with the same names as in `name` will be
-            overwritten. Otherwise only new components will be added and others will be
-            ignored.
-        kwargs : Any
-            Component attributes, e.g. x=[0.1, 0.2], can be list, pandas.Series
-            or pandas.DataFrame for time-varying
-
-        Returns
-        -------
-        new_names : pandas.index
-            Names of new components (including suffix)
-
-        Examples
-        --------
-        Add a single component:
-
-        >>> n = pypsa.Network()
-        >>> n.add("Bus", "my_bus_0")
-        Index(['my_bus_0'], dtype='object')
-        >>> n.add("Bus", "my_bus_1", v_nom=380)
-        Index(['my_bus_1'], dtype='object')
-        >>> n.add("Line", "my_line_name", bus0="my_bus_0", bus1="my_bus_1", length=34, r=2, x=4)
-        Index(['my_line_name'], dtype='object')
-
-        Add multiple components with static attributes:
-
-        >>> n.add("Load", ["load 1", "load 2"],
-        ...       bus=["1", "2"],
-        ...       p_set=np.random.rand(len(n.snapshots), 2))
-        Index(['load 1', 'load 2'], dtype='object')
-
-        Add multiple components with time-varying attributes:
-
-        >>> import pandas as pd, numpy as np
-        >>> buses = range(13)
-        >>> snapshots = range(7)
-        >>> n = pypsa.Network()
-        >>> n.set_snapshots(snapshots)
-        >>> n.add("Bus", buses)
-        Index(['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12'], dtype='object')
-        >>> # add load as numpy array
-        >>> n.add("Load",
-        ...       n.buses.index + " load",
-        ...       bus=buses,
-        ...       p_set=np.random.rand(len(snapshots), len(buses)))
-        Index(['0 load', '1 load', '2 load', '3 load', '4 load', '5 load', '6 load',
-               '7 load', '8 load', '9 load', '10 load', '11 load', '12 load'],
-              dtype='object', name='Bus')
-        >>> # add wind availability as pandas DataFrame
-        >>> wind = pd.DataFrame(np.random.rand(len(snapshots), len(buses)),
-        ...        index=n.snapshots,
-        ...        columns=buses)
-        >>> # use a suffix to avoid boilerplate to rename everything
-        >>> n.add("Generator",
-        ...       buses,
-        ...       suffix=' wind',
-        ...       bus=buses,
-        ...       p_nom_extendable=True,
-        ...       capital_cost=1e5,
-        ...       p_max_pu=wind)
-        Index(['0 wind', '1 wind', '2 wind', '3 wind', '4 wind', '5 wind', '6 wind',
-               '7 wind', '8 wind', '9 wind', '10 wind', '11 wind', '12 wind'],
-              dtype='object')
-
-
-        """
-        c = as_components(self, class_name)
-        # Process name/names to pandas.Index of strings and add suffix
-        single_component = np.isscalar(name)
-        names = pd.Index([name]) if single_component else pd.Index(name)
-        names = names.astype(str) + suffix
-
-        names_str = "name" if single_component else "names"
-        # Read kwargs into static and time-varying attributes
-        series = {}
-        static = {}
-
-        # Check if names are unique
-        if not names.is_unique:
-            msg = f"Names for {c.name} must be unique."
-            raise ValueError(msg)
-
-        for k, v in kwargs.items():
-            # If index/ columnes are passed (pd.DataFrame or pd.Series)
-            # - cast names index to string and add suffix
-            # - check if passed index/ columns align
-            msg = "{} has an index which does not align with the passed {}."
-            if isinstance(v, pd.Series) and single_component:
-                if not v.index.equals(self.snapshots):
-                    raise ValueError(msg.format(f"Series {k}", "network snapshots"))
-            elif isinstance(v, pd.Series):
-                # Cast names index to string + suffix
-                v = v.rename(
-                    index=lambda s: str(s)
-                    if str(s).endswith(suffix)
-                    else str(s) + suffix
-                )
-                if not v.index.equals(names):
-                    raise ValueError(msg.format(f"Series {k}", names_str))
-            if isinstance(v, pd.DataFrame):
-                # Cast names columns to string + suffix
-                v = v.rename(
-                    columns=lambda s: str(s)
-                    if str(s).endswith(suffix)
-                    else str(s) + suffix
-                )
-                if not v.index.equals(self.snapshots):
-                    raise ValueError(msg.format(f"DataFrame {k}", "network snapshots"))
-                if not v.columns.equals(names):
-                    raise ValueError(msg.format(f"DataFrame {k}", names_str))
-
-            # Convert list-like and 1-dim array to pandas.Series
-            if is_1d_list_like(v):
-                try:
-                    if single_component:
-                        v = pd.Series(v, index=self.snapshots)
-                    else:
-                        v = pd.Series(v)
-                        if len(v) == 1:
-                            v = v.iloc[0]
-                            logger.debug(
-                                "Single value sequence for %s is treated as a scalar "
-                                "and broadcasted to all components. It is recommended "
-                                "to explicitly pass a scalar instead.",
-                                k,
-                            )
-                        else:
-                            v.index = names
-                except ValueError as e:
-                    expec_str = (
-                        f"{len(self.snapshots)} for each snapshot."
-                        if single_component
-                        else f"{len(names)} for each component name."
-                    )
-                    msg = f"Data for {k} has length {len(v)} but expected {expec_str}"
-                    raise ValueError(msg) from e
-            # Convert 2-dim array to pandas.DataFrame
-            if isinstance(v, np.ndarray):
-                if v.shape == (len(self.snapshots), len(names)):
-                    v = pd.DataFrame(v, index=self.snapshots, columns=names)
-                else:
-                    msg = (
-                        f"Array {k} has shape {v.shape} but expected "
-                        f"({len(self.snapshots)}, {len(names)})."
-                    )
-                    raise ValueError(msg)
-
-            if isinstance(v, dict):
-                msg = (
-                    "Dictionaries are not supported as attribute values. Please use "
-                    "pandas.Series or pandas.DataFrame instead."
-                )
-                raise NotImplementedError(msg)
-
-            # Handle addition of single component
-            if single_component:
-                # Read 1-dim data as time-varying attribute
-                if isinstance(v, pd.Series):
-                    series[k] = pd.DataFrame(
-                        v.values, index=self.snapshots, columns=names
-                    )
-                # Read 0-dim data as static attribute
-                else:
-                    static[k] = v
-
-            # Handle addition of multiple components
-            elif not single_component:
-                # Read 2-dim data as time-varying attribute
-                if isinstance(v, pd.DataFrame):
-                    series[k] = v
-                # Read 1-dim data as static attribute
-                elif isinstance(v, pd.Series):
-                    static[k] = v.values
-                # Read scalar data as static attribute
-                else:
-                    static[k] = v
-
-        # Load static attributes as components
-        if static:
-            static_df = pd.DataFrame(static, index=names)
-        else:
-            static_df = pd.DataFrame(index=names)
-        _import_components_from_df(self, static_df, c.name, overwrite=overwrite)
-
-        # Load time-varying attributes as components
-        for k, v in series.items():
-            self._import_series_from_df(v, c.name, k, overwrite=overwrite)
-
-        return names
-
-    def remove(
-        self,
-        class_name: str,
-        name: str | int | Sequence[int | str],
-        suffix: str = "",
-    ) -> None:
-        """
-        Removes a single component or a list of components from the network.
-
-        Removes it from component DataFrames.
-
-        Parameters
-        ----------
-        class_name : str
-            Component class name
-        name : str, int, list-like or pandas.Index
-            Component name(s)
-        suffix : str, default=''
-            Suffix to be added to the component name(s)
-
-        Examples
-        --------
-        >>> n = pypsa.Network()
-        >>> n.snapshots = pd.date_range("2015-01-01", freq="h", periods=2)
-        >>> n.add("Bus", ["bus0", "bus1"])
-        Index(['bus0', 'bus1'], dtype='object')
-        >>> n.add("Bus", "bus2", p_min_pu=[1, 1])
-        Index(['bus2'], dtype='object', name='Bus')
-        >>> n.components.buses.static
-              v_nom type    x    y  ... v_mag_pu_max control generator  sub_network
-        Bus                         ...
-        bus0    1.0       0.0  0.0  ...          inf      PQ
-        bus1    1.0       0.0  0.0  ...          inf      PQ
-        bus2    1.0       0.0  0.0  ...          inf      PQ
-        <BLANKLINE>
-        [3 rows x 13 columns]
-
-        Remove a single component:
-        >>> n.remove("Bus", "bus2")
-
-
-        Any component data is dropped from the component DataFrames.
-        >>> n.components.buses.static
-            v_nom type    x    y  ... v_mag_pu_max control generator  sub_network
-        Bus                         ...
-        bus0    1.0       0.0  0.0  ...          inf      PQ
-        bus1    1.0       0.0  0.0  ...          inf      PQ
-        <BLANKLINE>
-        [2 rows x 13 columns]
-        >>> n.components.buses.dynamic.p_min_pu
-        Empty DataFrame
-        Columns: []
-        Index: [2015-01-01 00:00:00, 2015-01-01 01:00:00]
-
-        Remove multiple components:
-        >>> n.remove("Bus", ["bus0", "bus1"])
-
-        >>> n.components.buses.static
-        Empty DataFrame
-        Columns: [v_nom, type, x, y, carrier, unit, location, v_mag_pu_set, v_mag_pu_min, v_mag_pu_max, control, generator, sub_network]
-        Index: []
-
-        """
-        c = as_components(self, class_name)
-
-        # Process name/names to pandas.Index of strings and add suffix
-        names = pd.Index([name]) if np.isscalar(name) else pd.Index(name)
-        names = names.astype(str) + suffix
-
-        # Drop from static components
-        cls_static = self.static(c.name)
-        cls_static.drop(names, inplace=True)
-
-        # Drop from time-varying components
-        dynamic = self.dynamic(c.name)
-        for df in dynamic.values():
-            df.drop(df.columns.intersection(names), axis=1, inplace=True)
-
-    @deprecated(
-        deprecated_in="0.31",
-        removed_in="1.0",
-        details="Use `n.add` as a drop-in replacement instead.",
-    )
-    def madd(
-        self,
-        class_name: str,
-        names: Sequence,
-        suffix: str = "",
-        **kwargs: Any,
-    ) -> pd.Index:
-        """
-        Add multiple components to the network, along with their attributes.
-
-        .. deprecated:: 0.31
-          ``n.madd`` is deprecated and will be removed in a future version. Use
-            :py:meth:`pypsa.Network.add` instead. It can handle both single and multiple
-            removal of components.
-
-        Make sure when adding static attributes as pandas Series that they are indexed
-        by names. Make sure when adding time-varying attributes as pandas DataFrames that
-        their index is a superset of n.snapshots and their columns are a
-        subset of names.
-
-        Any attributes which are not specified will be given the default
-        value from :doc:`/user-guide/components`.
-
-        Parameters
-        ----------
-        class_name : string
-            Component class name in ("Bus", "Generator", "Load", "StorageUnit",
-            "Store", "ShuntImpedance", "Line", "Transformer", "Link").
-        names : list-like or pandas.Index
-            Component names
-        suffix : string, default ''
-            All components are named after names with this added suffix. It
-            is assumed that all Series and DataFrames are indexed by the original names.
-        kwargs
-            Component attributes, e.g. x=[0.1, 0.2], can be list, pandas.Series
-            of pandas.DataFrame for time-varying
-
-        """
-        return self.add(class_name=class_name, name=names, suffix=suffix, **kwargs)
-
-    @deprecated(
-        deprecated_in="0.31",
-        removed_in="1.0",
-        details="Use `n.remove` as a drop-in replacement instead.",
-    )
-    def mremove(self, class_name: str, names: Sequence) -> None:
-        """
-        Removes multiple components from the network.
-
-        .. deprecated:: 0.31
-          ``n.mremove`` is deprecated and will be removed in a future version. Use
-            :py:meth:`pypsa.Network.remove` instead. It can handle both single and multiple
-            removal of components.
-
-        ``n.mremove`` is deprecated and will be removed in version 1.0. Use
-        py:meth:`pypsa.Network.remove` instead. It can handle both single and multiple removal of
-        components.
-
-        Removes them from component DataFrames.
-
-        Parameters
-        ----------
-        class_name : string
-            Component class name
-        names : list-like
-            Component names
-
-        """
-        self.remove(class_name=class_name, name=names)
 
     def copy(
         self,
